test(ssr): move Radiant SSR writers into Node server suite
